Add check and confirm interaction for receipt email
In order to minimise emails with typos or incorrect, we are going to introduce a "check and confirm" interaction right after the applicant choose if they want to receive a copy of the submitted application.

This check and confirm follows the convention of Home Office (but we use our own styling to match the rest of the application):

https://home-office-digital-patterns.herokuapp.com/patterns/confirmation-loop

This check only appears when the user entered an email, as they can leave it blank, is not mandatory. Decision tree updated for this.