Improve the bypass functionality
It is now possible to answer one or more questions in the screener, and then do a bypass. These already answered questions will take precedence over the fixture data.

Also, now in the developer tools dropdown in the footer, the information about the court will appear, to quickly check where an application might be sent for a given postcode.