Reduce number of email alerts
As the service take up increase, there is also an increase in the number of emails we receive alerting us of failures in the applicant receipt email.

These failures are usually typos in the email address and nothing we can fix on our side.

In order to minimise these alerts so we only receive them when the error is genuine (or looks genuine) we are going to ignore the notify failures for the user receipt.
We still get them in the backoffice.