Ensure index uniqueness in two DB tables
These tables can only contain unique references to `C100Application`. There can't be duplicates.

There have been a rare occurrence where a duplicate entry was inserted (probably by submitting twice the form quickly, or some kind of race-condition difficult to replicate).

In order to avoid this from happening again, and as we do with other tables already, we are going to enforce the uniqueness of the index `c100_application_id` column.