Add one time task to fix some rows in audit table
The `completed_applications_audit` table was created a few months after the release of the service.

Although we tried to backtrack as many applications as possible from the early days, the consequence was for some (1045) applications we couldn't reliably assign their reference code, leaving this field as `nil` in these rows.

This is not a very smart decision as the `reference_code` is the primary key and any mass update is done using this PK, which means if we want to update a bunch of records without reference code, we must use the exact timestamp which is not ideal.

Also, some of these old rows have been mistakenly re-assigned a different court, meaning the analytics are incorrect.

This one-time task will, based on an old backup, re-assign the proper court to these 1045 records, and at the same time assign also a `reference_code` so they are no longer `nil` (even if this code means nothing).