Add a basic in-memory cache for court slugs
Instead of using memoization of the `court_data`, we replace it with a very basic `ActiveSupport::Cache::MemoryStore` which in essence does the same but allow us to set an expiration per-key and also it is easier to test.

Also in the near future, this should be replace with a `RedisCacheStore` as otherwise it is suboptimal, having more than one instance, or loosing the cache on deploys.