Adapt some tests to new Rails version
After the bump to Rails 5.2.4 and also due to the removal of some old framework defaults (we are now using the defaults for this version of rails) a few tests had to be adapted.

Also, disable pg_stat_statements extension

This was added by default to the DB long time ago when we run the service on Heroku, and thus the database dump kept this plugin enabled despite not having their corresponding "enable" migration on Rails.

The side effect was, every time we run a migration, the plugin got added to the schema.

We are going to disable it for good with a proper migration.