Add voicemail consent radios to applicant
In the applicant contact details, we are going to ask a new question, whether they consent the court to leave a voicemail. This is a Yes/No question.

Updated the view to add this radio, as well as reorder the fields in both, applicant and respondent, moving the email field to the top.