Refactor the failed emails report notification
Previously this email was sent to a single recipient address, declared on an ENV variable.

Now that we have a functional back office, with the ability to retry these failed emails, we can send the notification to all the back office admin users so they are aware and can trigger the email resend themselves.

Still an ENV variable needs to be set in order for these emails to be sent, otherwise they will be disabled.