Replace declaration checkbox with full name input
In preparation for solicitor's journey, we are replacing the declaration single checkbox in the Check Your Answers page, with a full name input where the signee (whether an applicant, or a solicitor) can enter their name.

If the applicant has a solicitor, they also will be able to select who is the signee (applicant or solicitor). If they don't have a solicitor (they answered NO to that question) then we don't show these radio options and default to `applicant` as the signee capacity.

Some minor updates to the PDF so we can print these new details, and also make sure it works backwards for compatibility with completed applications still in database.