Remove legal representative question from screener
In preparation for the solicitor's journey release, we are removing this question from the screener, as it will now be possible to have a legal representative when applying online.

Removed the DB attribute, locales, views, routes, etc. Updated the screener decision tree.

Also, some copy changes, including removing the solicitor bullet from the criteria in the home page, and updating the "entrypoint" and "what is needed" pages.