Add gem and configuration to use Sidekiq
In preparation to start using a persistent job queue and processor.

We are going to use Sidekiq as it integrates transparently with Redis that we are already using, and also let us continue using our current ActiveJob classes without changes.

Updated `docker-compose` to include sidekiq (remember we DO NOT use docker-compose on production).

A new ENV variable `QUEUE_ADAPTER` can be used to change the adapter. The default if not present is `sidekiq` but on development machines is useful to be able to change it to `async`.