Add the application version to the audit record
This will help when creating statistics, as some fields or questions were introduced in different versions of the application and thus it is not expected old audit records will contain all metadata fields.

As of today, current version is 4, saved applications (drafts) are in 3, and old audit records might be in any version from 1 to 3. We are not going to backdate these old audit records as would be difficult to pinpoint the exact version and it is not needed for current statistics.