Stop using ministryofjustice/docker-templates
The Dockerfiles in the repo [docker-templates](https://github.com/ministryofjustice/docker-templates) is being deprecated and it is recommended to keep Dockerfiles and docker images as lightweight and simple as possible.

In this case we are going to refactor the Dockerfile to use an alpine image and add dependencies on top.

This service is a bit special as it requires some extra dependencies in order for the PDF generation to work fine and with nice fonts.

These PDF gems had to be updated to work with the new Alpine-based image.