Refactor status endpoint
Make it a bit more useful including Sidekiq checks, and also remove
the dependency we had with Court Tribunal Finder because it is just
too unreliable and we don't have any control over it so even if it
goes down, we should not consider our service unhealthy.

Still, this endpoint should not be used for kubernetes probes.
Use `ping` instead.