Create `court_arrangements` table and model
This new table will be used to create the new fields needed for the redesign of the special measures steps, beginning with the language step.

Once all the special measures steps have been migrated, we will be able to remove the old DB fields from the main `c100_applications` table.