Add language step in new `attending_court` namespace
The main objective here will be to replicate the existing special measures steps, but in a separate namespace so they can be independent from each other and let us introduce new fields.

For a while these 2 journeys will have to cohabit together until all new applications are migrated to the new version.

We are not enabling yet in this PR the new journey.