Change order of the litigation capacity step
We are moving this question one step back, so it goes before the language question, instead of after, like it is now.

This will help with the reorganising of the rest of the attending court questions. In essence we maintain this question as it is, and we will not use the new table for these fields, at least for now.

The PDF order does not change.