Duplicate intermediary step in new namespace
As we did with language we are creating (in this case an identical) copy of the already existing `intermediary` step, but in the new namespace `AttendingCourt`, and using the DB fields from the associated table `court_arrangement`.

Some minor copy changes along the way as per design.

This is not yet wired to the main journey, but can be accessed knowing the URL.

Next PR will add the neccessary CYA and PDF changes.

Also, for future needs, a couple of fields for Welsh language have been added to the `court_arrangements` table. We are not using them at the moment.