Prepare litigation capacity for later migration
So we don't have to duplicate the form objects, controllers, views, etc. as well as changes to CYA and PDF for litigator capacity questions, we are goint to take a pragmatic approach where we continue using the existing code, but copy over the values to the new database table.

When the time comes to delete the old code and attributes for the old attending court journey, it will be much easier just to start using the new table as values for existing applications will be present.