Update PDF additional information section
This needs to cope with the old version of the language question, and the new version.

New version works with check boxes so we need to see if at least one, no matter which one, was selected.

Once all applications are migrated to the new version, we can cleanup the code.