Refactor completions audit log
This is growing too much with the addition of the special arrangements metadata.

Refactor a bit the code to split the metadata to a separate helper class.

Added a new metadata attribute (`arrangements`) to store a collection (array) of arrangement strings based on what options the applicant chose, either yes-no or check boxes.
Only for new applications using the new `court_arrangements` table. Older applications will just store an empty array.