Bump to ruby 2.6.5 and some other gems
We are not yet bumping to Ruby 2.7 but that will come next in a few weeks. It is still a bit new and there might be some incompatibilities so better to wait.

Ruby 2.6.5 is the last release in the 2.6 branch.

Other minor bumps to some other gems, with no breaking changes.

Minor changes to Dockerfile and docker-compose to cope with some of these changes. Specifically, `xz-libs` seems to be needed (outside of the build-deps) in order for a shared library `liblzma.so.5`, used by nokogiri, to work.