Fix long standing wrong naming (other parties)
We were not careful when we designed the URLs for people of kind `OtherParty` and we pluralised them, against the convention, making it difficult to work with their URLs.

Corrected this by renaming `other_parties` namespace to `other_party` (as we have with `applicant`, `respondent`...)

This allow us to cleanup the patch we introduced in `PersonUrlHelper`.

It looks like a lot of files have changed but it is mostly relocating them into their correct namespace.