Rake task to find and enqueue a court data refresh
Self explanatory.

This can be run ad-hoc as a quick and dirty way of updating some stale data when CTF changes something on their side and then leaves us with applications being sent to the incorrect email address or things like that (Newcastle rings a bell...)

It should not be used for anything else and probably we would never need to use it more than once in a while.