Allow resending emails in back office, even not failed ones
There is a requirement to be able to resend recent emails that were successfully sent but for whatever reason, they need to be resend.

In particular, due to an IT outage, some emails where sent without error, but some courts did not receive them.
But can also happen when an applicant or a court says they've not received anything (or they deleted the email by mistake).

This expose a new functionality in the back office so, given a reference code, it is possible to trigger again the sending of emails (court or applicant) at any time, during the 28 days from the creation of the application.