Add an application fulfilment custom validator
This validator will run only on completion of the application (still not wired up, that will come up in a follow-up PR) and will check some preconditions are met in order for the application to be considered "valid" and to be submitted to court. Essentially a quick sanity check.

This is not intended to be an in-deep validation of every single attribute (as that already happens at the step level, one by one as the applicant progress through the application) but more for cases where the user jumped some steps or deleted by mistake all applicants, or all respondents, etc.

Still to do in a follow-up PR the wiring to the CYA page, and the presentation of the errors using the `FulfilmentErrorsPresenter` class and the `check_your_answers/fulfilment_error` partial (not included in this PR as we are awaiting design).

Note: as there is nothing yet on the service to "see" this working, apart from the tests, this can also be tested quickly by going to the rails console and running:

```ruby
a = C100Application.new
a.valid?(:completion)
FulfilmentErrorsPresenter.new(a).errors
```