Present fulfilment errors in CYA page
Follow-up to PR #818.

Show the errors, if any, in the CYA page. Still design is TBD so the partials are just placeholders, showing basic details about the error.

As part of this PR and to protect this new feature from being released to production before is ready, added a simple mechanism to feature-flag the validations: it will run ONLY if we are using fake court data (a fixture, like we do when ue the bypass functionality on local/staging envs).

This way, on production there is no risk of this validation being run just yet.

Note: it is possible to avoid the validations also on local/staging if you start the screener and answer at least one question, then you can use the bypass functionality (or just continue all the way to the end answering questions) and the fulfilment validator will not trigger.