Add validation for children and respondents
Same as we did with applicants, we add validation for having at least one child and at least one respondent.

For now this will be all. We can decide to include more validations later on.

Reworked a bit the loop to retrieve the error details so it works when there is more than one, like we have now.

The final little PR after this one will be to enable this for all users (without feature-flag) to be able to release to production.