Customise email copy when applicant is under age
Expose a helper method on the `Applicants` collection to query if there is at least one applicant under age (returning true or false).

This method is later used on Notify emails, but can also be used for the confirmation pages or any other place where we need to know if the application as a whole contains at least one under age.

Send the yes/no to Notify to be able to show/hide the instructions when applicant is under age in the confirmation email.

Tweaked a bit the card payment instructions as per suggestions from content designer.
The under18s instructions are part of the Notify template.