Add under age data to the audit table
In order to accurately track the number of applications considered 'under age' (containing at least 1 applicant under 18s), we add this data to the audit table, as we've done with other details.

Also show this in the back office when retrieving the details of an application.

The default for old applications (lacking this data) is false as we've had the screener question all this time and didn't allow under 18s.