Migrate about pages to new design system
Migrate all pages linked to from the footer (or header like the "Report a problem" link) to the new desig system.

This is not a search and replace (sadly). It is a time consuming task because we must replace old markup with appropiate new one, as well as classes, but also use headings consistently to create a clear hierarchy throughout the service.

The headings mark up must be semantic as per guidelines here:
https://design-system.service.gov.uk/styles/typography/#headings

Also removed obsolete markup and replaced `strong` tags with either headers or proper font weight classes.