Migrate error summary to use new form builder
For a while we will need to maintain both error summary so the new one is named `govuk_error_summary` and will be used in the correct position in the markup of the view as it must be inside the main container, and inside the grid.

As this change requires updating all the views, for now we maintain compatibility with the old error summary that will be rendered outside the main content as before.