Update markup of screener email consent question
And update the features.

TODO: the gem does not automatically localise the radio labels. It needs to be passed as `label: { text: 'foobar' }` which is kind of annoying.
Probably we can raise a PR to improve this.

This and other issues will be discovered as we advance with the migration.