First two forms in journey migrated
These are the two first questions after the screener.

Note during the migration, as it advances, new issues can be found. I found the changes done to the forked gem to cope with localisations for radios was not covering all scenarios.

Some more changes have been made to the forked gem and we shall see if this covers everything we might need in this service.

From now on some steps will be quite complex so it could be we need to make some more amendments (trying as much as possible to always do this in a backwards compatible way so perhaps we can integrate these improvements in the upstream repo).