Migrate attending court steps to new design system
I've jumped ahead to almost the end of the journey in order to migrate this whole block of 4 questions involving mostly check boxes, as this let us test the gem in this scenario, with some check boxes revealing text areas.

Dramatically simplified the `ArrangementsCheckBoxesForm` concern as instead of separate attributes for each check box, we now use a single attribute to contain the selected check boxes.
This is so it works better with the new gem too.