Updated Aldershot court
The old page is no longer valid but there is a new slug for this court, so adding it to the service.

This fixes the failing smoke test.

Note: I've added to the .gitignore the `node_modules` as we are now using it (on develop branch) and could be commited by mistake when switching between branches.