Change court slugs from whitelist to blacklist
Instead of having a big list of all courts available to use the online service, we are now moving to a black list approach where only those courts that do NOT want to take part, or receive applications, will be added.

As we are now nationwide in all courts, this should not have any impact but will avoid the constant issues where some courts modify their "slug" without notifying us. We will not need to maintain a list anymore, win-win.

Removed the smoke tests as now they are not needed. We can enable smoke test in the future if we want, to test other things so I will leave the config on CircleCI just in case but commented out.