Add helper to create links with button shape
This is a very common use case and we are using it in many places (more to come with the migration).

The helper will output the expected markup and will simplify if we later on need to change the classes or something as ideally there will only be one place to touch.

It tries to be smart about overriding (or not) some attributes, to cope with most common combinations, but later on we could improve it if we need more complex use-cases.

Updated already migrated pages to use this helper.