Migrate first date form element (MIAM certification date)
This PR introduces the migration of the first date element found in the application, MIAM certification date (a couple other pages have been migrated to get there but it was just radio options and copy).

For the dates previously we were using a custom date gem `gov_uk_date_fields` that had the ability not only to produce the markup but also to send parameters and process this parameters.

Now, the new gem does not provide that functionality, it only renders the view element, so anything in the backend is up to us. Also, it sends the parameter in the Rails way for multi params (appending a suffix to the parameter name).

Still we are going to leave the old gem because the steps not yet migrated will continue using it but once we've migrated all the dates in the application we can remove the gem and some other code.