Migrate the abuse concerns journey
Most of the pages here are a loop, meaning we reuse again and again the same "view" with the same form components but we just tweak a bit the copy based on a variable.

This journey is triggered if at least one of the safety questions is answered as YES (for example, substance abuse).

Refactored the custom stylesheet for the `SideBar` to simplify them and make it clear these are custom (have the `app-` prefix)